+++ tests/cases/resolve-changelog/normalizeEmptyLines.js
@@ -1,0 +1,25 @@
+const emptyLines = `
+  foo
+ 
+bar  
+   
+
+`;
+
+const emptyLinesResult = `
+  foo
+
+bar  
+
+
+`;
+
+module.exports = {
+	cases: [
+		{
+			description: 'Normalize empty lines',
+			input: emptyLines,
+			output: emptyLinesResult
+		}
+	]
+};